ccl: move jwthelper to pkg/security/jwthelper
Informs #164337
Release note: None
