+++ pkg/BUILD.bazel
@@ -1064,7 +1064,6 @@ GO_TARGETS = [
     "//pkg/ccl/schemachangerccl:schemachangerccl",
     "//pkg/ccl/schemachangerccl:schemachangerccl_test",
     "//pkg/ccl/securityccl/fipsccl:fipsccl",
-    "//pkg/ccl/securityccl/jwthelper:jwthelper",
     "//pkg/ccl/serverccl/adminccl:adminccl_test",
     "//pkg/ccl/serverccl/diagnosticsccl:diagnosticsccl_test",
     "//pkg/ccl/serverccl/statusccl:statusccl_test",
@@ -1962,6 +1961,7 @@ GO_TARGETS = [
     "//pkg/security/distinguishedname:distinguishedname",
     "//pkg/security/dyncert:dyncert",
     "//pkg/security/dyncert:dyncert_test",
+    "//pkg/security/jwthelper:jwthelper",
     "//pkg/security/password:password",
     "//pkg/security/password:password_test",
     "//pkg/security/pprompt:pprompt",
