+++ pkg/ccl/jwtauthccl/BUILD.bazel
@@ -10,7 +10,7 @@ go_library(
     importpath = "github.com/cockroachdb/cockroach/pkg/ccl/jwtauthccl",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/ccl/securityccl/jwthelper",
+        "//pkg/security/jwthelper",
         "//pkg/security/username",
         "//pkg/server/telemetry",
         "//pkg/settings",
