+++ pkg/security/jwthelper/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "jwthelper",
     srcs = ["groups.go"],
-    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/securityccl/jwthelper",
+    importpath = "github.com/cockroachdb/cockroach/pkg/security/jwthelper",
     visibility = ["//visibility:public"],
     deps = [
         "@com_github_cockroachdb_errors//:errors",
